Revert "Provide information about which files changed in the results hash"
This reverts commit 9412a1a0f033babe15963685f3b3fff878ad0faf.

Conflicts:
	tests/index.js

Unfortunately, this does not have the desired outcome. Broccoli uses FS
watching merely as an indicator that it should re-stat and potentially
rebuild. In some scenarios this works, but in the general case it will
likely be problematic.

I believe the correct solution is to derive what ultimately changed by
diffing the old and new outputs.